requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
openpyxl>=3.0.10
python-docx>=0.8.11
tqdm>=4.64.0
lxml>=4.9.0
